wrltemplate.bbclass: Rework the process ordering
Process ordering will run from the bbpath forward, unless a 'startpath' is
defined.  If that is defined it will only run from the startpath to the end.

This change synced up the behavior of prior Wind River Linux template
processing and this new version.

image.bbclass change:

At present 'require' enforces the file must be there and 'blank' files are not
ignored as they are with 'inherit'.  We provide two dummy files that do nothing
to satisfy this requirement.

Signed-off-by: Mark Hatle <mark.hatle@windriver.com>