wrtemplate.bbclass: Fix a template require problem
It would have been possible, but unlikely, if two layers had the same
template defined to end up with both of them via a 'require'.  This fixes
the problem by only allowing the 'continue' (keep looking) in when
'firstfound' is False.  It's only set to 'False' in a self-referenced
'require' situation.

Signed-off-by: Mark Hatle <mark.hatle@windriver.com>