wrltemplate.bbclass: Move the includes into a conf file
It is too late to include templates that change USER_CLASSES or other
variables that affect load order in the 'wrltemplate.bbclass'.  Instead we
intercept the bitbake.conf, load the default AND then load our templates.

We're late enough in the process we can override default settings, but early
enough that inherits and classes have not yet been processed.

Signed-off-by: Mark Hatle <mark.hatle@windriver.com>