wrltemplate.bbclass: Copy in the template contents
Issue: LIN9-3600 LIN9-3358

Before we would use require to bring in template contents, however if the
system configuration changed it could cause a failure that was only recoverable
via a rm -f conf/wr*.

This new way copies in the contents so no matter what we can always parse the
config.  However, this introduces a problem where if the template changes we
may not detect it.  So we need to add code that verified that the template
mtime has not changed.

This is needed for Yocto Project compatible, as the compatibility checking
script will run different configurations in a single pass.  So we needed to
support this workflow without requiring the manual 'rm' step.

Signed-off-by: Mark Hatle <mark.hatle@windriver.com>