Refactor for recent OpenEmbedded Changes (no functional change)
Contents of the bitbake.conf were moved to the wrltemplate.bbclass (which is
always loaded by the layer.conf USER_CLASSES).

Moving the code to the USER_CLASSES ensures that it is always loaded.  The
load order will still allow the user to override the defaults.

The former image.bbclass was renamed to wrltemplate_image.bbclass and modified
to load the image include files d.getVar instead of ${...}.  The behavior of
the system has changed that if undefined will trigger a parse error, even
though the value was not used.  The layer.conf was also modified to add the
wrltemplate_image as an IMAGE_CLASSES in all cases.

Using the IMAGE_CLASSES avoids having to intercept and replace the existing
image.bbclass.  This could fail if the BBPATH order changed.

Tests have been updated for the latest OE version and the README file was
updated to match.  All existing tests have been run to show no functional
difference.

Signed-off-by: Mark Hatle <mark.hatle@windriver.com>