wrtemplate.bbclass: Add WRTEMPLATE_CLASSES as a way for templates to add classes
Issue: LIN10-2141

If a template changes a variable that affects INHERIT, it will not take effect
because it is too late to extend INHERIT.  Inherit is only executed once,
prior to loading of the wrtemplate.bbclass.

We detect this situation and issue an error.  This will ensure that the user
does not create or run a template that does not do what is expected.

We also add a new WRTEMPLATE_CLASSES variable, which a template can use to
load additional classes to the system.

Signed-off-by: Mark Hatle <mark.hatle@windriver.com>